organize files in components
