+++ app/components/main-content/LandingPageMainHero.tsx
